backend/requirements.txt
Django>=4.2
djangorestframework
drf-yasg
django-cors-headers
django-environ
psycopg2-binary
redis
gunicorn
whitenoise
requests
